refactor: split app schedule runtime
